+++ include/rocksdb/table.h
@@ -121,31 +121,39 @@ struct PlainTablePropertyNames {
   static const std::string kEncodingType;
 };
 
-// -- Plain Table with prefix-only seek
-// For this factory, you need to set Options.prefix_extrator properly to make it
-// work. Look-up will starts with prefix hash lookup for key prefix. Inside the
-// hash bucket found, a binary search is executed for hash conflicts. Finally,
-// a linear search is used.
+const uint32_t kPlainTableVariableLength = 0;
+
+struct PlainTableOptions {
 // @user_key_len: plain table has optimization for fix-sized keys, which can be
 //                specified via user_key_len.  Alternatively, you can pass
 //                `kPlainTableVariableLength` if your keys have variable
 //                lengths.
+uint32_t user_key_len = kPlainTableVariableLength;
+
 // @bloom_bits_per_key: the number of bits used for bloom filer per prefix. You
 //                      may disable it by passing a zero.
+int bloom_bits_per_key = 10;
+
 // @hash_table_ratio: the desired utilization of the hash table used for prefix
 //                    hashing. hash_table_ratio = number of prefixes / #buckets
 //                    in the hash table
+double hash_table_ratio = 0.75;
+
 // @index_sparseness: inside each prefix, need to build one index record for how
 //                    many keys for binary search inside each hash bucket.
 //                    For encoding type kPrefix, the value will be used when
 //                    writing to determine an interval to rewrite the full key.
 //                    It will also be used as a suggestion and satisfied when
 //                    possible.
+size_t index_sparseness = 16;
+
 // @huge_page_tlb_size: if <=0, allocate hash indexes and blooms from malloc.
 //                      Otherwise from huge page TLB. The user needs to reserve
 //                      huge pages for it to be allocated, like:
 //                          sysctl -w vm.nr_hugepages=20
 //                      See linux doc Documentation/vm/hugetlbpage.txt
+size_t huge_page_tlb_size = 0;
+
 // @encoding_type: how to encode the keys. See enum EncodingType above for
 //                 the choices. The value will determine how to encode keys
 //                 when writing to a new SST file. This value will be stored
@@ -153,13 +161,21 @@ struct PlainTablePropertyNames {
 //                 file, which makes it possible for users to choose different
 //                 encoding type when reopening a DB. Files with different
 //                 encoding types can co-exist in the same DB and can be read.
+EncodingType encoding_type = kPlain;
 
-const uint32_t kPlainTableVariableLength = 0;
-extern TableFactory* NewPlainTableFactory(
-    uint32_t user_key_len = kPlainTableVariableLength,
-    int bloom_bits_per_prefix = 10, double hash_table_ratio = 0.75,
-    size_t index_sparseness = 16, size_t huge_page_tlb_size = 0,
-    EncodingType encoding_type = kPlain, bool full_scan_mode = false);
+// @full_scan_mode: mode for reading the whole file one record by one without
+//                  using the index.
+  bool full_scan_mode = false;
+};
+
+// -- Plain Table with prefix-only seek
+// For this factory, you need to set Options.prefix_extrator properly to make it
+// work. Look-up will starts with prefix hash lookup for key prefix. Inside the
+// hash bucket found, a binary search is executed for hash conflicts. Finally,
+// a linear search is used.
+
+extern TableFactory* NewPlainTableFactory(const PlainTableOptions& options =
+                                              PlainTableOptions());
 
 #endif  // ROCKSDB_LITE