+++ HISTORY.md
@@ -6,6 +6,7 @@
 
 ### Public API changes
 * DBOptions.db_paths now is a vector of a DBPath structure which indicates both of path and target size
+* NewPlainTableFactory instead of bunch of parameters now accepts PlainTableOptions, which is defined in include/rocksdb/table.h
 
 
 ## 3.3.0 (7/10/2014)